introduce cl-quil.si package, eliminate :: refs in Quilt
Eliminate use of package names with double colons (::) in Quilt, by
two methods: (1) exporting many semi-internal symbols to a new
package, cl-quil.si, and referring to them in that package; and (2) by
referring to symbols already exported (from CL-Quil) with a single
colon.

The new package, cl-quil.si (or quil.si, for short) is intended to
export system internal (SI) symbols, that is, some symbols not
generally intended to be used by the wider public, but more intended
for "friends" -- packages or facilities that are prepared to go a bit
below the surface level API.

References to most such symbols in Quilt, which previously referred to
quil::<symbol>, are changed in this commit to instead refer to
quil.si:<symbol>. (Quil.si is a nickname for cl-quil.si.)

In addition, the following symbols were already being exported from
the cl-quil package, so they were simply changed to be referred to
with single colon.

  TOK
  TOKEN-PAYLOAD
  TOKEN-TYPE
  IS-CONSTANT
  APPLICATION-ARGUMENTS
  APPLICATION-OPERATOR
  GATE-DEFINITION
  PARSED-PROGRAM-EXECUTABLE-CODE
  QUBIT-INDEX
  INSTRUCTION-FMT
  PARSE-QUIL
  GATE-DEFINITION-TO-GATE
  PERMUTATION-GATE
  PERMUTATION-GATE-PERMUTATION